Consolidate `FHighlightsTags` and `FEarlyAccess` to a single callback feature on store pages
