+++ src/js/Content/Features/Store/Storefront/CStoreFront.js
@@ -1,16 +1,15 @@
 import ContextType from "../../../Modules/Context/ContextType";
 import {CStoreBase} from "../Common/CStoreBase";
-import FMonitorStoreFront from "./FMonitorStoreFront";
-import FHomePageTab from "./FHomePageTab";
 import FCustomizer from "../Common/FCustomizer";
+import FHomePageTab from "./FHomePageTab";
 
 export class CStoreFront extends CStoreBase {
 
     constructor() {
+
         super(ContextType.STORE_FRONT, [
-            FMonitorStoreFront,
-            FHomePageTab,
             FCustomizer,
+            FHomePageTab,
         ]);
     }
 }
